Add some element for team course and bulletin board section
