+++ app/src/main/java/com/nltv/chafenqi/model/team/TeamBasicInfo.kt
@@ -55,9 +55,9 @@ data class TeamBasicInfo(
             promotable = false,
             lastMonthActivityPoints = 0,
             currentActivityPoints = 0,
-            courseTrack1String = "",
-            courseTrack2String = "",
-            courseTrack3String = "",
+            courseTrack1String = "0,0",
+            courseTrack2String = "0,0",
+            courseTrack3String = "0,0",
             courseHealth = 0,
             courseLastModifiedAt = 0,
             pinnedMessageId = null,
