+++ app/src/main/java/com/nltv/chafenqi/view/home/team/HomeTeamPage.kt
@@ -20,6 +20,7 @@ import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.AddComment
 import androidx.compose.material.icons.filled.GroupRemove
 import androidx.compose.material.icons.filled.Warning
 import androidx.compose.material.icons.outlined.Warning
@@ -29,7 +30,10 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExtendedFloatingActionButton
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
@@ -78,7 +82,6 @@ import kotlinx.datetime.Clock
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun HomeTeamPage(navController: NavController) {
-    val coroutinesScope = rememberCoroutineScope()
     val model: HomeTeamPageViewModel = viewModel()
     val state by model.uiState.collectAsStateWithLifecycle()
 
@@ -103,7 +106,7 @@ fun HomeTeamPage(navController: NavController) {
     Scaffold(
         topBar = {
             CenterAlignedTopAppBar(
-                title = { Text(text = "团队") },
+                title = { Text(text = state.team.info.displayName) },
                 scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(),
                 colors = TopAppBarDefaults.topAppBarColors(
                     containerColor = MaterialTheme.colorScheme.surface,
@@ -118,6 +121,15 @@ fun HomeTeamPage(navController: NavController) {
                     }
                 }
             )
+        },
+        floatingActionButton = {
+            if (pagerState.currentPage == 3) {
+                ExtendedFloatingActionButton(
+                    text = { Text("发布留言") },
+                    icon = { Icon(Icons.Default.AddComment, contentDescription = "发布留言") },
+                    onClick = {}
+                )
+            }
         }
     ) { paddingValues ->
         Column(
@@ -146,192 +158,13 @@ fun HomeTeamPage(navController: NavController) {
                 modifier = Modifier.fillMaxSize()
             ) { index ->
                 when (index) {
-                    0 -> { HomeTeamPageMemberList(list = state.team.members) }
+                    0 -> { HomeTeamPageMemberList() }
                     1 -> {}
-                    2 -> {}
-                    3 -> {}
+                    2 -> { HomeTeamPageCourseSection() }
+                    3 -> { HomeTeamPageBulletinBoardSection() }
                 }
             }
         }
-    }
-}
-
-@Composable
-fun HomeTeamPageMemberList(
-    list: List<TeamMember>
-) {
-    val haptics = LocalHapticFeedback.current
-    val model: HomeTeamPageViewModel = viewModel()
-    val state by model.uiState.collectAsStateWithLifecycle()
-
-    var selectedTeamMemberUserId by rememberSaveable {
-        mutableLongStateOf(-1L)
-    }
-
-    LazyColumn (
-        modifier = Modifier.fillMaxSize()
-    ) {
-        items(
-            count = list.size,
-            key = { index -> index }
-        ) {
-            HomeTeamPageMemberEntry(
-                member = list[it],
-                onLongClick = {
-                    haptics.performHapticFeedback(HapticFeedbackType.LongPress)
-                    selectedTeamMemberUserId = list[it].userId
-                }
-            )
-        }
-    }
-
-    if (selectedTeamMemberUserId != -1L) {
-        val member = state.team.members.firstOrNull { it.userId == selectedTeamMemberUserId }
-        if (member != null) {
-            HomeTeamPageMemberManageSheet(
-                member = member,
-            ) {
-                selectedTeamMemberUserId = -1L
-            }
-        } else {
-            selectedTeamMemberUserId = -1L
-        }
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun HomeTeamPageMemberEntry(
-    member: TeamMember,
-    onLongClick: () -> Unit
-) {
-    val context = LocalContext.current
-
-    @Composable
-    fun MemberInfoRow(
-        description: String,
-        value: String
-    ) {
-        Row (
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text(text = description, style = MaterialTheme.typography.bodySmall)
-            Text(text = value, style = MaterialTheme.typography.bodySmall)
-        }
-    }
-
-    Card (
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(10.dp)
-            .combinedClickable(
-                onClick = {},
-                onLongClick = {
-                    onLongClick()
-                },
-                onLongClickLabel = "管理成员"
-            ),
-        shape = ShapeDefaults.Medium,
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = 4.dp,
-            pressedElevation = 6.dp
-        )
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(10.dp)
-                .padding(vertical = 2.dp)
-        ) {
-            Row (
-                horizontalArrangement = Arrangement.SpaceBetween,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(85.dp)
-                    .padding(bottom = 10.dp)
-            ) {
-                Column(
-                    modifier = Modifier
-                        .weight(1f)
-                        .fillMaxHeight(),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.SpaceBetween
-                ) {
-                    Text(text = member.trophy, style = MaterialTheme.typography.titleSmall, color = Color.Gray)
-                    Text(text = member.nickname, fontWeight = FontWeight.Bold, style = MaterialTheme.typography.titleMedium)
-
-                    Row (
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Row (
-                            horizontalArrangement = Arrangement.spacedBy(10.dp)
-                        ) {
-                            Text("Rating", style = MaterialTheme.typography.titleSmall)
-                            Text(member.rating, fontWeight = FontWeight.Bold, style = MaterialTheme.typography.titleSmall)
-                        }
-
-                        Row (
-                            horizontalArrangement = Arrangement.spacedBy(10.dp)
-                        ) {
-                            Text("游玩次数", style = MaterialTheme.typography.titleSmall)
-                            Text("${member.playCount}", fontWeight = FontWeight.Bold, style = MaterialTheme.typography.titleSmall)
-                        }
-                    }
-                }
-                AsyncImage(
-                    model = member.avatar,
-                    contentDescription = "头像",
-                    modifier = Modifier
-                        .padding(start = 8.dp)
-                        .size(75.dp)
-                        .clip(RoundedCornerShape(12.dp)),
-                    contentScale = ContentScale.FillHeight,
-                    clipToBounds = true
-                )
-            }
-
-            MemberInfoRow(description = "加入时间：", value = member.joinAt.toDateString(context))
-            MemberInfoRow(description = "贡献点数：", value = "${member.activityPoints}")
-            MemberInfoRow(description = "最后游玩时间：", value = member.lastActivityAt.toDateString(context))
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun HomeTeamPageMemberManageSheet(
-    member: TeamMember,
-    onDismissRequest: () -> Unit
-) {
-    var shouldShowConfirmDialog by rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    ModalBottomSheet(
-        onDismissRequest = onDismissRequest
-    ) {
-        ListItem(
-            headlineContent = { Text(text = "移除该成员") },
-            leadingContent = { Icon(Icons.Default.GroupRemove, contentDescription = "移除成员") },
-            modifier = Modifier.clickable(
-                onClick = {
-                    shouldShowConfirmDialog = true
-                }
-            )
-        )
-    }
-
-    if (shouldShowConfirmDialog) {
-        HomeTeamPageConfirmDialog(
-            icon = Icons.Outlined.Warning,
-            title = "确认移除",
-            message = "您确定要将此成员从您的团队中移除吗？该操作无法撤销。",
-            onConfirm = { shouldShowConfirmDialog = false },
-            onDismissRequest = { shouldShowConfirmDialog = false }
-        )
     }
 }
 
@@ -359,25 +192,5 @@ fun HomeTeamPageConfirmDialog(
                 Text("取消")
             }
         }
-    )
-}
-
-@Preview
-@Composable
-fun HomeTeamPageMemberEntryPreview() {
-    HomeTeamPageMemberEntry(
-        member = TeamMember(
-            id = 1,
-            userId = 1L,
-            nickname = "LOUISE/",
-            joinAt = Clock.System.now().epochSeconds,
-            avatar = "https://new.chunithm-net.com/chuni-mobile/html/mobile/img/f9ed64ced3d22730.png",
-            trophy = "What's Up? Pop!",
-            rating = "17.03",
-            activityPoints = 80240L,
-            playCount = 215,
-            lastActivityAt = Clock.System.now().epochSeconds,
-        ),
-        onLongClick = {}
     )
 }
