+++ app/src/main/java/com/nltv/chafenqi/view/home/team/HomeTeamPageBulletinBoardSection.kt
@@ -1,0 +1,13 @@
+package com.nltv.chafenqi.view.home.team
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AddComment
+import androidx.compose.material3.ExtendedFloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HomeTeamPageBulletinBoardSection() {
+
+}
