+++ app/version.properties
@@ -1,5 +1,5 @@
-#Fri Dec 06 22:44:02 JST 2024
-VERSION_BUILD=1075
+#Sat Dec 07 00:13:34 JST 2024
+VERSION_BUILD=1089
 VERSION_MAJOR=1
 VERSION_MINOR=1
 VERSION_PATCH=9
